Move kernels outside of modules
