+++ PartialImplementations/PartialImplementations.jl
@@ -2,18 +2,16 @@ using RRTMGP
 using DocStringExtensions
 using RRTMGP.mo_optical_props
 using RRTMGP.mo_simple_netcdf
-using RRTMGP.mo_rte_solver_kernels
 using RRTMGP.fortran_intrinsics
 using RRTMGP.mo_util_array
 using RRTMGP.mo_gas_optics_rrtmgp
 using RRTMGP.mo_gas_concentrations
-using RRTMGP.mo_rte_lw
+using RRTMGP.RTESolver
 using RRTMGP.mo_fluxes
 using RRTMGP.mo_load_coefficients
 using RRTMGP.mo_rfmip_io
 using RRTMGP.mo_source_functions
 using RRTMGP.mo_cloud_optics
-using RRTMGP.mo_rte_sw
 using RRTMGP.mo_load_cloud_coefficients
 
 import RRTMGP.mo_optical_props: validate!, subset_range!, delta_scale!, get_nmom
