+++ docs/src/RTE/RTESolver.md
@@ -1,13 +1,13 @@
 # RTE Solver
 
 ```@meta
-CurrentModule = RRTMGP
+CurrentModule = RRTMGP.RTESolver
 ```
 
 ```@docs
-mo_rte_sw.rte_sw!
-mo_rte_sw.expand_and_transpose
-mo_rte_lw.rte_lw!
+rte_sw!
+expand_and_transpose
+rte_lw!
 ```
