+++ src/RRTMGP.jl
@@ -5,16 +5,13 @@ include("fortran_intrinsics.jl")
 include("Utilities.jl")
 
 include(joinpath("rte","mo_util_array.jl"))
-include(joinpath("rte","kernels","mo_fluxes_broadband_kernels.jl"))
-include(joinpath("rte","kernels","mo_rte_solver_kernels.jl"))
 include(joinpath("rte","mo_optical_props.jl"))
 include(joinpath("rte","mo_fluxes.jl"))
 include(joinpath("rte","mo_source_functions.jl"))
 include(joinpath("extensions","mo_fluxes_byband_kernels.jl"))
 include(joinpath("extensions", "cloud_optics", "mo_cloud_optics.jl"))
 # include(joinpath("extensions","mo_fluxes_byband.jl"))
-include(joinpath("rte","mo_rte_sw.jl"))
-include(joinpath("rte","mo_rte_lw.jl"))
+include(joinpath("rte","RTESolver.jl"))
 
 include(joinpath("rrtmgp","mo_rrtmgp_constants.jl"))
 include(joinpath("rrtmgp","mo_gas_concentrations.jl"))
