+++ test/allsky.jl
@@ -6,18 +6,16 @@ const to = TimerOutput()
 using NCDatasets
 using RRTMGP.mo_optical_props
 using RRTMGP.mo_simple_netcdf
-using RRTMGP.mo_rte_solver_kernels
 using RRTMGP.fortran_intrinsics
 using RRTMGP.mo_util_array
 using RRTMGP.mo_gas_optics_rrtmgp
 using RRTMGP.mo_gas_concentrations
-using RRTMGP.mo_rte_lw
+using RRTMGP.RTESolver
 using RRTMGP.mo_fluxes
 using RRTMGP.mo_load_coefficients
 using RRTMGP.mo_rfmip_io
 using RRTMGP.mo_source_functions
 using RRTMGP.mo_cloud_optics
-using RRTMGP.mo_rte_sw
 using RRTMGP.mo_load_cloud_coefficients
 
 include("mo_cloud_sampling.jl")
