+++ test/rfmip_clear_sky_lw.jl
@@ -2,12 +2,11 @@ using Test
 using RRTMGP
 using NCDatasets
 using RRTMGP.mo_optical_props
-using RRTMGP.mo_rte_solver_kernels
 using RRTMGP.fortran_intrinsics
 using RRTMGP.mo_util_array
 using RRTMGP.mo_gas_optics_rrtmgp
 using RRTMGP.mo_gas_concentrations
-using RRTMGP.mo_rte_lw
+using RRTMGP.RTESolver
 using RRTMGP.mo_fluxes
 using RRTMGP.mo_load_coefficients
 using RRTMGP.mo_rfmip_io
